Add AsTimeSeries template (not working now). For #1 change SQLite config to the most strict.
Explicit commit vs implicit looks like reduces the Packer race a lot. But it's still there sometimes. Which makes me think the issue is composite.
There could be something with the SLState. The error is so weird currently that bottom up from the primitives parts is the way to go.

Also currently using experimental SQLite branch, but the theoretically useful features from that branch are actually useless or unmeasurable with
the current load, so should use stable SQLite. At least with that build + most strict concurrency setting there will be no more questions, since
SQLite in the most reliable software out there.