Disable block prefetch. #1
It was written in 15 minutes as a proof of concept and indeed significantly improves read throughput.
However, ConcurrentDictionary is not a good choice because it's not always thread safe / atomic, and
the implementation was dropping buffers. Need special careful work for prefetch, and ensure that
additional synchronization doesn't kill prefetch benefits.

Also need to decide if prefetch is an implementation detail of a storage or should be done at cursor
level for any storage.